organize app directory
