+++ decorators/trimmer_decorator.rb
