+++ models/nameable.rb
