+++ models/student.rb
