+++ models/teacher.rb
